+++ zerver/lib/event_schema.py
@@ -1753,6 +1753,7 @@ update_message_flags_remove_event = event_dict_type(
                     ],
                     optional_keys=[
                         ("mentioned", bool),
+                        ("alerted", bool),
                         ("user_ids", ListType(int)),
                         ("stream_id", int),
                         ("topic", str),